Written missing test and renamed a file.
Written a missing test for format_dict. Renamed netbox_data to match the
class name.
