+++ pynetbox_data_uploader/lib/enums/dcim_device_id.py
