+++ pynetbox_data_uploader/tests/test_netbox_data.py
